Update bug report instructions
The instructions for creating a bug report have been updated to include a request for screenshots. Now, users are prompted not only to recreate their actions and post the output, but also, if possible, to provide visual evidence of their workflow. This additional information will help us better identify and resolve the reported bugs.